Moved entities into their respective packages.
