+++ wingit/src/main/java/com/G2T5203/wingit/WingitApplication.java
@@ -1,16 +1,11 @@
 package com.G2T5203.wingit;
 
-import com.G2T5203.wingit.entities.WingitUser;
+import com.G2T5203.wingit.user.WingitUser;
 import com.G2T5203.wingit.user.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Profile;
-import org.springframework.core.env.Environment;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.core.io.support.PropertiesLoaderUtils;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -19,9 +14,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import java.io.IOException;
 import java.time.LocalDate;
-import java.util.Arrays;
 import java.util.Properties;
-import java.util.zip.InflaterOutputStream;
 
 @SpringBootApplication
 public class WingitApplication {
