+++ wingit/src/main/java/com/G2T5203/wingit/booking/BookingRepository.java
@@ -1,7 +1,6 @@
 package com.G2T5203.wingit.booking;
 
-import com.G2T5203.wingit.entities.Booking;
-import com.G2T5203.wingit.entities.RouteListingPk;
+import com.G2T5203.wingit.routeListing.RouteListingPk;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
