+++ wingit/src/main/java/com/G2T5203/wingit/plane/Plane.java
@@ -1,4 +1,4 @@
-package com.G2T5203.wingit.entities;
+package com.G2T5203.wingit.plane;
 
 import jakarta.persistence.*;
