+++ wingit/src/main/java/com/G2T5203/wingit/plane/PlaneController.java
@@ -1,13 +1,7 @@
 package com.G2T5203.wingit.plane;
 
-import com.G2T5203.wingit.entities.Plane;
-import com.G2T5203.wingit.user.UserController;
 import jakarta.validation.Valid;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.transaction.UnexpectedRollbackException;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
