+++ wingit/src/main/java/com/G2T5203/wingit/plane/PlaneRepository.java
@@ -1,6 +1,5 @@
 package com.G2T5203.wingit.plane;
 
-import com.G2T5203.wingit.entities.Plane;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PlaneRepository extends JpaRepository<Plane, String> {
