+++ wingit/src/main/java/com/G2T5203/wingit/route/RouteRepository.java
@@ -1,6 +1,5 @@
 package com.G2T5203.wingit.route;
 
-import com.G2T5203.wingit.entities.Route;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
