+++ wingit/src/main/java/com/G2T5203/wingit/routeListing/RouteListingController.java
@@ -1,6 +1,5 @@
 package com.G2T5203.wingit.routeListing;
 
-import com.G2T5203.wingit.entities.RouteListing;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
