+++ wingit/src/main/java/com/G2T5203/wingit/routeListing/RouteListingRepository.java
@@ -1,7 +1,5 @@
 package com.G2T5203.wingit.routeListing;
 
-import com.G2T5203.wingit.entities.RouteListing;
-import com.G2T5203.wingit.entities.RouteListingPk;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
