+++ wingit/src/main/java/com/G2T5203/wingit/seat/SeatController.java
@@ -1,6 +1,5 @@
 package com.G2T5203.wingit.seat;
 
-import com.G2T5203.wingit.entities.Seat;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
