+++ wingit/src/main/java/com/G2T5203/wingit/seat/SeatPk.java
@@ -1,5 +1,6 @@
-package com.G2T5203.wingit.entities;
+package com.G2T5203.wingit.seat;
 
+import com.G2T5203.wingit.plane.Plane;
 import jakarta.persistence.Embeddable;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
