+++ wingit/src/main/java/com/G2T5203/wingit/seatListing/SeatListing.java
@@ -1,5 +1,6 @@
-package com.G2T5203.wingit.entities;
+package com.G2T5203.wingit.seatListing;
 
+import com.G2T5203.wingit.booking.Booking;
 import jakarta.persistence.*;
 
 @Entity
