+++ wingit/src/main/java/com/G2T5203/wingit/seatListing/SeatListingNotFoundException.java
@@ -1,6 +1,5 @@
 package com.G2T5203.wingit.seatListing;
 
-import com.G2T5203.wingit.entities.SeatListingPk;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
