+++ wingit/src/main/java/com/G2T5203/wingit/seatListing/SeatListingSimpleJson.java
@@ -1,7 +1,5 @@
 package com.G2T5203.wingit.seatListing;
 
-import com.G2T5203.wingit.entities.SeatListing;
-
 import java.time.LocalDateTime;
 
 public class SeatListingSimpleJson {
