+++ wingit/src/main/java/com/G2T5203/wingit/user/CustomAuthorityDeserializer.java
@@ -1,4 +1,4 @@
-package com.G2T5203.wingit.entities;
+package com.G2T5203.wingit.user;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
