+++ wingit/src/main/java/com/G2T5203/wingit/user/UserRepository.java
@@ -1,6 +1,5 @@
 package com.G2T5203.wingit.user;
 
-import com.G2T5203.wingit.entities.WingitUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
