+++ wingit/src/main/java/com/G2T5203/wingit/user/UserService.java
@@ -1,13 +1,10 @@
 package com.G2T5203.wingit.user;
 
-import com.G2T5203.wingit.entities.WingitUser;
 import jakarta.transaction.Transactional;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import javax.swing.text.html.Option;
-import java.awt.*;
 import java.util.List;
 import java.util.Optional;
