+++ wingit/src/test/java/com/G2T5203/wingit/plane/PlaneControllerTest.java
@@ -1,7 +1,6 @@
 package com.G2T5203.wingit.plane;
 
 import com.G2T5203.wingit.TestUtils;
-import com.G2T5203.wingit.entities.Plane;
 import com.G2T5203.wingit.user.UserRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
